requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
uvicorn
sentence-transformers
faiss-cpu
numpy
transformers
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.1
pinecone
google-generativeai 
pinecone-client
torch